Update tests to reflect design change
As I noodled on it, it didn't seem to make much sense to me to write
something which can detect the presence of a numeric input to validate
it is numeric, and something separate to extract that numeric input from
arbitrary text.

By definition, it should only return the valid doubles found in the
string passed, and null otherwise.