Update number extraction regex, provide commentary
This was a doozy. At some point, I knew everything I was looking up was
wildly out of scope, but I had tripped over curiosity.
Something-something, premature optimization, YAGNI, etc.

I'm curious what an implementation of that would look like, to build a
regex to extract a number from a string based on what the NumberStyle is
set to. I was a little surprised that by googling I couldn't find
anything relevant.