Add XML comments
Normally I'm in favor of only commenting to explain why something
happens when I'm dealing with Private or otherwise internal members. If
they aren't beinge exposed outside the class, "how" and "what" are only
relevant from within the class. Can't hurt, but it feels like a waste of
time on non-public members.