Add TODO for book.cs
In the middle of rewriting other services, I'm realizing, my data model
is weird. Why the hell was I implementing this all with an int book ID
instead of letting the machine handle it with a List<book>? It makes
everything I'm doing so much more complicated. Was I thinking about
ISBNs? Was I being cheeky that way? Idk, but this is genuinely not that
useful so I might rip it out and see how my thinking changes.