Implement DI pattern for BookService
Keeping the BookService static was technically working, and less
complex, but couldn't be usefully tested against. On the one hand it
meant that I didn't have to worry about singletons or other weirdness,
on the other hand, it was not an effective expression of the state of
the program.

However, an API controller is apparently instantiated just about every
time a call is made. The API controller is, similarly, stateless. So
there needs to be *some* way to maintain a handle on the state.

Marking BookService as a singleton would've been similarly useful for
managing state, but similarly difficult to test, and fragile.

The DI pattern, even though it involves more indirection than feels
necessary, makes the BookService class testable by letting me treat it
as just another class, and managing a single instance of it through the
IServicesCollection functionality.