Add dependency-injection correctly
Dependency injection here relies on the resource being injected and the
resource it is inject into having a similar lifecycle. That is,
DbContext is scoped, so that it is invoked or handled each time a
request is made. It cannot be given to a singleton, which will live
for the runtime of the program.

By adding the static qualifier to BookService's bookCollection list, I'm
allowing it to have shared state for the lifecycle of the program. This
is only really necessary as I'm learning and refactoring.