Fix a confusing http return edgecase
The assumption that returned index would by definition *not* be the same
index submitted means that if the duplicate entry happened to be on the
same index, I would erroneously be returning a 404 instead of a 409.

Maybe I should find some other way to express that, like 200 OK, but I
mean for duplicate entries to have the semantics of a conflict, even if
the duplicate entry is on the same index.