FIX bug with Sqlite vs in-memory == eval
I don't know what the source of the bug is, probably that it's trying to
run equality on the Sqlite engine directly, rather than pulling the
value in locally and running it with the notnet assemblies.

What a nightmare. Everything was hunky-dory in my tests, but failed in
the live environment.